refactor: move validators service to shared projectversionservice
