+++ api/zanata-common-api/src/main/java/org/zanata/rest/service/ProjectVersionResource.java
@@ -276,4 +276,27 @@ public interface ProjectVersionResource {
             @PathParam("docId") String docId,
             @DefaultValue("en-US") @PathParam("localeId") String localeId,
             FilterFields filterFields);
+
+    /**
+     * Returns Validation settings for a Project Version.
+     *
+     * @param projectSlug project identifier
+     *
+     * @param versionSlug project version identifier
+     *
+     * @return The following response status codes will be returned from this
+     *         operation:<br>
+     *         OK(200) - Containing the Project data.<br>
+     *         NOT FOUND(404) - If a Project could not be found for the given
+     *         parameters.<br>
+     *         INTERNAL SERVER ERROR(500) - If there is an unexpected error in
+     *         the server while performing this operation.
+     */
+
+    @GET
+    @Produces({ MediaType.APPLICATION_JSON })
+    @Path(VERSION_SLUG_TEMPLATE + "/validators")
+    public Response getValidationSettings(
+            @PathParam("projectSlug") String projectSlug,
+            @PathParam("versionSlug") String versionSlug);
 }
