+++ server/zanata-frontend/src/app/editor/actions/settings-actions.js
@@ -40,7 +40,7 @@ export const fetchSettings = () => dispatch => dispatch({
  */
 export function fetchValidationSettings (dispatch, projectSlug, versionSlug) {
   const validationSettingsUrl =
-    `${apiUrl}/project/validators/project/${projectSlug}/version/${versionSlug}`
+    `${apiUrl}/project/${projectSlug}/version/${versionSlug}/validators`
   dispatch({
     [CALL_API_ENHANCED]: {
       endpoint: validationSettingsUrl,
